requirements.txt
python-telegram-bot==20.8
torch==2.3.0
torchvision==0.18.0
diffusers==0.29.2
transformers==4.44.2
accelerate==0.33.0
pillow
opencv-python-headless
numpy